Increased max-width of the word-limit CSS rule
Changed .word-limit-{max-width:40em;} to .word-limit-{max-width:45em;}
to increase the readability of the paragraphs.

And here is the why of it:

To quote a passage from “Web Style Guide – Basic design principles for
creating web sites”.

The ideal line length for text layout is based on the physiology of the
human eye… At normal reading distance the arc of the visual field is
only a few inches – about the width of a well-designed column of text,
or about 12 words per line. Research shows that reading slows and
retention rates fall as line length begins to exceed the ideal width,
because the reader then needs to use the muscles of the eye and neck to
track from the end of one line to the beginning of the next line. If
the eye must traverse great distances on the page, the reader is easily
lost and must hunt for the beginning of the next line. Quantitative
studies show that moderate line lengths significantly increase the
legibility of text.

Web Style Guide – Basic Design Principles for Creating Website
Patrick J. Lynch and Sarah Horton
2nd edition, page 97.