Added edit and delete buttons to answers
Edit and delete buttons will only be visible to the content author. So
the original creator of the answer will be able to edit and delete
(including comments against their answer). Comment authors on the
answers will have the ability to edit and delete their own comments but
not the answer.